added build.sh, and activate command
